+++ CHANGELOG.md
@@ -7,11 +7,11 @@
   - [ ] Visual indication that packages has update
 - [ ] Uninstall package
   - [ ] Light bulb in `packman.json`
-- [ ] Remove all package artifacts from project
 - [x] UI for package search
   - [x] Preview of what's being installed
   - [x] Select exactly which files to install
   - [x] Optional creation of package folders
+  - [x] Optional saving to `packman.json`
   - [x] Preselect latest stable version
 - [x] JSON manifest Intellisense
   - [x] JSON Schema